+++ bluetooth_low_energy_platform_interface/lib/src/bluetooth_low_energy_peer.dart
@@ -1,7 +1,0 @@
-import 'uuid.dart';
-
-/// An object that represents a remote device.
-abstract class BluetoothLowEnergyPeer {
-  /// The UUID associated with the peer.
-  UUID get uuid;
-}