+++ bluetooth_low_energy_platform_interface/lib/src/central.dart
@@ -1,4 +1,0 @@
-import 'bluetooth_low_energy_peer.dart';
-
-/// A remote device connected to a local app, which is acting as a peripheral.
-abstract class Central extends BluetoothLowEnergyPeer {}