+++ bluetooth_low_energy_platform_interface/lib/src/central_manager.dart
@@ -1,28 +1,20 @@
 import 'dart:typed_data';
 
-import 'package:log_service/log_service.dart';
+import 'package:hybrid_core/hybrid_core.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
-import 'bluetooth_low_energy_manager.dart';
-import 'central_event_args.dart';
-import 'gatt_characteristic.dart';
-import 'gatt_characteristic_write_type.dart';
-import 'gatt_descriptor.dart';
-import 'gatt_service.dart';
-import 'peripheral.dart';
+import 'common.dart';
 
-/// An object that scans for, discovers, connects to, and manages peripherals.
-///
-/// Platform-specific implementations should implement this class to support [CentralManager].
-abstract class CentralManager extends PlatformInterface
+/// Platform-specific implementations should implement this class to support [CentralManagerImpl].
+abstract base class CentralManagerImpl extends PlatformInterface
     with LoggerProvider, LoggerController
-    implements BluetoothLowEnergyManager {
+    implements CentralManager {
   static final Object _token = Object();
 
-  static CentralManager? _instance;
+  static CentralManagerImpl? _instance;
 
-  /// The default instance of [CentralManager] to use.
-  static CentralManager get instance {
+  /// The default instance of [CentralManagerImpl] to use.
+  static CentralManagerImpl get instance {
     final instance = _instance;
     if (instance == null) {
       throw UnimplementedError(
@@ -32,16 +24,19 @@ abstract class CentralManager extends PlatformInterface
   }
 
   /// Platform-specific implementations should set this with their own
-  /// platform-specific class that extends [CentralManager] when
+  /// platform-specific class that extends [CentralManagerImpl] when
   /// they register themselves.
-  static set instance(CentralManager instance) {
+  static set instance(CentralManagerImpl instance) {
     PlatformInterface.verifyToken(instance, _token);
     _instance = instance;
   }
 
-  /// Constructs a [CentralManager].
-  CentralManager() : super(token: _token);
+  /// Constructs a [CentralManagerImpl].
+  CentralManagerImpl() : super(token: _token);
+}
 
+/// An object that scans for, discovers, connects to, and manages peripherals.
+abstract interface class CentralManager implements BluetoothLowEnergyManager {
   /// Tells the central manager discovered a peripheral while scanning for devices.
   Stream<DiscoveredEventArgs> get discovered;
 
@@ -52,10 +47,19 @@ abstract class CentralManager extends PlatformInterface
   Stream<GattCharacteristicNotifiedEventArgs> get characteristicNotified;
 
   /// Scans for peripherals that are advertising services.
-  Future<void> startDiscovery();
+  ///
+  /// The [serviceUUIDs] argument is an array of [UUID] objects that the app is
+  /// interested in. Each [UUID] object represents the [UUID] of a service that
+  /// a peripheral advertises.
+  Future<void> startDiscovery({
+    List<UUID>? serviceUUIDs,
+  });
 
   /// Asks the central manager to stop scanning for peripherals.
   Future<void> stopDiscovery();
+
+  /// Returns a list of the peripherals connected to the system.
+  Future<List<Peripheral>> retrieveConnectedPeripherals();
 
   /// Establishes a local connection to a peripheral.
   Future<void> connect(Peripheral peripheral);
@@ -95,4 +99,67 @@ abstract class CentralManager extends PlatformInterface
     GattDescriptor descriptor, {
     required Uint8List value,
   });
+}
+
+/// A remote peripheral device.
+abstract interface class Peripheral implements BluetoothLowEnergyPeer {}
+
+base class PeripheralImpl extends BluetoothLowEnergyPeerImpl
+    implements Peripheral {
+  PeripheralImpl({
+    required super.uuid,
+  });
+
+  @override
+  int get hashCode => uuid.hashCode;
+
+  @override
+  bool operator ==(Object other) {
+    return other is Peripheral && other.uuid == uuid;
+  }
+}
+
+/// The discovered event arguments.
+base class DiscoveredEventArgs extends EventArgs {
+  /// The disvered peripheral.
+  final Peripheral peripheral;
+
+  /// The rssi of the peripheral.
+  final int rssi;
+
+  /// The advertisement of the peripheral.
+  final Advertisement advertisement;
+
+  /// Constructs a [DiscoveredEventArgs].
+  DiscoveredEventArgs(this.peripheral, this.rssi, this.advertisement);
+}
+
+/// The connection state cahnged event arguments.
+base class ConnectionStateChangedEventArgs extends EventArgs {
+  /// The peripheral which connection state changed.
+  final Peripheral peripheral;
+
+  /// The connection state.
+  final bool connectionState;
+
+  /// Constructs a [ConnectionStateChangedEventArgs].
+  ConnectionStateChangedEventArgs(
+    this.peripheral,
+    this.connectionState,
+  );
+}
+
+/// The GATT characteristic notified event arguments.
+base class GattCharacteristicNotifiedEventArgs extends EventArgs {
+  /// The GATT characteristic which notified.
+  final GattCharacteristic characteristic;
+
+  /// The notified value.
+  final Uint8List value;
+
+  /// Constructs a [GattCharacteristicNotifiedEventArgs].
+  GattCharacteristicNotifiedEventArgs(
+    this.characteristic,
+    this.value,
+  );
 }