+++ bluetooth_low_energy_platform_interface/lib/src/event_args.dart
@@ -1,2 +1,0 @@
-/// The base event arguments.
-abstract class EventArgs {}