+++ bluetooth_low_energy_platform_interface/lib/src/gatt_characteristic_write_type.dart
@@ -1,9 +1,0 @@
-/// The write type for a GATT characteristic.
-enum GattCharacteristicWriteType {
-  // Write with response
-  withResponse,
-  // Write without response
-  withoutResponse,
-  // Write with response and waiting for confirmation
-  // reliable,
-}