+++ bluetooth_low_energy_platform_interface/lib/src/peripheral.dart
@@ -1,4 +1,0 @@
-import 'bluetooth_low_energy_peer.dart';
-
-/// A remote peripheral device.
-abstract class Peripheral extends BluetoothLowEnergyPeer {}