+++ bluetooth_low_energy_platform_interface/lib/src/peripheral_manager.dart
@@ -1,27 +1,20 @@
 import 'dart:typed_data';
 
-import 'package:log_service/log_service.dart';
+import 'package:hybrid_core/hybrid_core.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
-import 'advertisement.dart';
-import 'bluetooth_low_energy_manager.dart';
-import 'central.dart';
-import 'gatt_characteristic.dart';
-import 'gatt_service.dart';
-import 'peripheral_event_args.dart';
+import 'common.dart';
 
-/// An object that manages and advertises peripheral services exposed by this app.
-///
-/// Platform-specific implementations should implement this class to support [PeripheralManager].
-abstract class PeripheralManager extends PlatformInterface
+/// Platform-specific implementations should implement this class to support [PeripheralManagerImpl].
+abstract base class PeripheralManagerImpl extends PlatformInterface
     with LoggerProvider, LoggerController
-    implements BluetoothLowEnergyManager {
+    implements PeripheralManager {
   static final Object _token = Object();
 
-  static PeripheralManager? _instance;
+  static PeripheralManagerImpl? _instance;
 
-  /// The default instance of [PeripheralManager] to use.
-  static PeripheralManager get instance {
+  /// The default instance of [PeripheralManagerImpl] to use.
+  static PeripheralManagerImpl get instance {
     final instance = _instance;
     if (instance == null) {
       throw UnimplementedError(
@@ -31,16 +24,20 @@ abstract class PeripheralManager extends PlatformInterface
   }
 
   /// Platform-specific implementations should set this with their own
-  /// platform-specific class that extends [PeripheralManager] when
+  /// platform-specific class that extends [PeripheralManagerImpl] when
   /// they register themselves.
-  static set instance(PeripheralManager instance) {
+  static set instance(PeripheralManagerImpl instance) {
     PlatformInterface.verifyToken(instance, _token);
     _instance = instance;
   }
 
-  /// Constructs a [PeripheralManager].
-  PeripheralManager() : super(token: _token);
+  /// Constructs a [PeripheralManagerImpl].
+  PeripheralManagerImpl() : super(token: _token);
+}
 
+/// An object that manages and advertises peripheral services exposed by this app.
+abstract interface class PeripheralManager
+    implements BluetoothLowEnergyManager {
   /// Tells that the local peripheral device received an Attribute Protocol (ATT) read request for a characteristic with a dynamic value.
   Stream<GattCharacteristicReadEventArgs> get characteristicRead;
 
@@ -77,4 +74,78 @@ abstract class PeripheralManager extends PlatformInterface
     required Uint8List value,
     Central? central,
   });
+}
+
+/// A remote device connected to a local app, which is acting as a peripheral.
+abstract interface class Central implements BluetoothLowEnergyPeer {}
+
+base class CentralImpl extends BluetoothLowEnergyPeerImpl implements Central {
+  CentralImpl({
+    required super.uuid,
+  });
+
+  @override
+  int get hashCode => uuid.hashCode;
+
+  @override
+  bool operator ==(Object other) {
+    return other is Central && other.uuid == uuid;
+  }
+}
+
+/// The GATT characteristic written event arguments.
+class GattCharacteristicReadEventArgs {
+  /// The central which read this characteristic.
+  final Central central;
+
+  /// The GATT characteristic which value is read.
+  final GattCharacteristic characteristic;
+
+  /// The value.
+  final Uint8List value;
+
+  /// Constructs a [GattCharacteristicReadEventArgs].
+  GattCharacteristicReadEventArgs(
+    this.central,
+    this.characteristic,
+    this.value,
+  );
+}
+
+/// The GATT characteristic written event arguments.
+class GattCharacteristicWrittenEventArgs {
+  /// The central which wrote this characteristic.
+  final Central central;
+
+  /// The GATT characteristic which value is written.
+  final GattCharacteristic characteristic;
+
+  /// The value.
+  final Uint8List value;
+
+  /// Constructs a [GattCharacteristicWrittenEventArgs].
+  GattCharacteristicWrittenEventArgs(
+    this.central,
+    this.characteristic,
+    this.value,
+  );
+}
+
+/// The GATT characteristic notify state changed event arguments.
+class GattCharacteristicNotifyStateChangedEventArgs {
+  /// The central which set this notify state.
+  final Central central;
+
+  /// The GATT characteristic which notify state changed.
+  final GattCharacteristic characteristic;
+
+  /// The notify state.
+  final bool state;
+
+  /// Constructs a [GattCharacteristicNotifyStateChangedEventArgs].
+  GattCharacteristicNotifyStateChangedEventArgs(
+    this.central,
+    this.characteristic,
+    this.state,
+  );
 }