+++ bluetooth_low_energy_windows/lib/bluetooth_low_energy_windows.dart
@@ -4,6 +4,6 @@ import 'src/my_central_manager.dart';
 
 abstract class BluetoothLowEnergyWindows {
   static void registerWith() {
-    CentralManager.instance = MyCentralManager();
+    CentralManagerAPI.instance = MyCentralManager();
   }
 }