+++ bluetooth_low_energy_windows/lib/src/my_central_manager.dart
@@ -7,7 +7,7 @@ import 'my_api.dart';
 import 'my_gatt_characteristic2.dart';
 import 'my_gatt_descriptor2.dart';
 
-class MyCentralManager extends CentralManager
+class MyCentralManager extends CentralManagerAPI
     implements MyCentralManagerFlutterApi {
   final MyCentralManagerHostApi _api;
   final StreamController<BluetoothLowEnergyStateChangedEventArgs>